KokkosBatched - perftest of lu
