+++ src/batched/KokkosBatched_LU_Serial_Internal.hpp
@@ -73,10 +73,11 @@ namespace KokkosBatched {
       enum : int {
         mbAlgo = Algo::LU::Blocked::mb<Kokkos::Impl::ActiveExecutionMemorySpace>()
       };
-      const typename Kokkos::Details::ArithTraits<ValueType>::mag_type one(1.0), minus_one(-1.0);          
+      const typename MagnitudeScalarType<ValueType>::type one(1.0), minus_one(-1.0);
+
       const int k = (m < n ? m : n);
       if (k <= 0) return 0;
-
+      
       InnerLU<mbAlgo> lu(as0, as1);
           
       InnerTrsmLeftLowerUnitDiag<mbAlgo>    trsm_llu(as0, as1, as0, as1);
